pyproject.toml
[tool.poetry]
name = "testovoe"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10"
fastapi = {extras = ["standard"], version = "^0.115.0"}
elasticsearch = "^8.15.1"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.5.2"
uvicorn = "^0.30.6"
chardet = "^5.2.0"
aiohttp = "^3.10.5"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.35"}
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
